requirements.txt
ckiptagger==0.2.1
tensorflow==2.8.2
numpy==1.22.4
pandas==1.4.2
pymongo==4.1.1
dnspython==2.2.1
scikit-learn==1.1.1
openpyxl==3.0.10
google-api-python-client==2.51.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2